Fix adding instances
